+++ mex/common/models/extracted_data.py
@@ -1,16 +1,21 @@
-from typing import TYPE_CHECKING, Annotated, Any
+from typing import Annotated, TypeVar
 
-from pydantic import Field, model_validator, validate_call
+from pydantic import Field
 
 from mex.common.models.entity import BaseEntity
 from mex.common.types import (
+    ExtractedIdentifier,
     ExtractedPrimarySourceIdentifier,
+    MergedIdentifier,
     MergedPrimarySourceIdentifier,
 )
 
 MEX_PRIMARY_SOURCE_IDENTIFIER = ExtractedPrimarySourceIdentifier("00000000000001")
 MEX_PRIMARY_SOURCE_IDENTIFIER_IN_PRIMARY_SOURCE = "mex"
 MEX_PRIMARY_SOURCE_STABLE_TARGET_ID = MergedPrimarySourceIdentifier("00000000000000")
+
+_MergedIdentifierT = TypeVar("_MergedIdentifierT", bound=MergedIdentifier)
+_ExtractedIdentifierT = TypeVar("_ExtractedIdentifierT", bound=ExtractedIdentifier)
 
 
 class ExtractedData(BaseEntity):
@@ -25,20 +30,6 @@ class ExtractedData(BaseEntity):
     This class also adds a validator to automatically set identifiers for provenance.
     See below, for a full description.
     """
-
-    if TYPE_CHECKING:  # pragma: no cover
-        # Sometimes multiple primary sources describe the same activity, resource, etc.
-        # and a complete metadata item can only be created by merging these fragments.
-        # The `stableTargetId` is part of all extracted models to allow MEx to identify
-        # which items describe the same thing and should be merged to create a complete
-        # metadata item. The name `stableTargetId` might be a bit misleading, because
-        # the "stability" is only guaranteed for one "real world" or "digital world"
-        # thing having the same ID in MEx over time. But it is not a guarantee, that the
-        # same metadata sources contribute to the complete metadata item. The naming has
-        # its historical reasons, but we plan to change it in the near future.
-        # Because we anticipate that items have to be merged, the `stableTargetId` is
-        # also used as the foreign key for all fields containing references.
-        stableTargetId: Any
 
     hadPrimarySource: Annotated[
         MergedPrimarySourceIdentifier,
@@ -73,127 +64,40 @@ class ExtractedData(BaseEntity):
         ),
     ]
 
-    # TODO make stable_target_id and identifier computed fields (MX-1435)
-    @model_validator(mode="before")
-    @classmethod
-    @validate_call
-    def set_identifiers(cls, values: dict[str, Any]) -> dict[str, Any]:  # noqa: C901
-        """Ensure identifiers and provenance attributes are set for this instance.
-
-        All extracted data classes have four important identifiers that are defined
-        by `BaseEntity`, `ExtractedData` and the concrete classes themselves.
-
-        - identifierInPrimarySource
-        - hadPrimarySource
-        - identifier
-        - stableTargetId
-
-        Every time we create a new instance of an extracted item, we automatically
-        validate that these identifiers are set correctly.
-
-        We check that `identifierInPrimarySource` and `hadPrimarySource` are provided,
-        because otherwise we cannot reliably determine the origin of this item.
-        These two identifiers are the only two that need to be set during extraction.
-
-        Next, we query the configured `IdentityProvider` to determine whether this item
-        already has an `identifier` and `stableTargetId`. If not, we let the identity
-        provider generate new identifiers.
-
-        If an `identifier` has been passed to the constructor, we check that it matches
-        with what we got from the identity provider, because we don't allow any system
-        to change the association from `identifierInPrimarySource` and
-        `hadPrimarySource` to the `identifier`.
-        A use-case for passing a matching `identifier` to the constructor would be
-        parsing an already extracted item from an NDJSON file or an API endpoint.
-
-        If a `stableTargetId` has been passed to the constructor, we use that as the
-        new value, because changes to the stable target ID are generally allowed.
-        A use-case for changing the `stableTargetId` will be the matching of
-        multiple extracted items (see `BaseEntity.stableTargetId` for details).
+    def _get_identifier(
+        self, identifier_type: type[_ExtractedIdentifierT]
+    ) -> _ExtractedIdentifierT:
+        """Consult the identity provider to get the `identifier` for this item.
 
         Args:
-            values: Raw values to validate
-
-        Raises:
-            ValueError: If `identifier` was supplied but does not match the id provider
-            ValueError: If `identifierInPrimarySource` was missing
-            ValueError: If `hadPrimarySource` was missing
+            identifier_type: ExtractedIdentifier-subclass to cast the identifier to
 
         Returns:
-            Values with identifier and provenance attributes
+            Identifier of the correct type
         """
-        # break import cycle, sigh
-        from mex.common.identity import get_provider
+        from mex.common.identity import get_provider  # break import cycle, sigh
 
-        # validate ID in primary source and primary source ID
-        if identifier_in_primary_source := values.get("identifierInPrimarySource"):
-            if isinstance(identifier_in_primary_source, list):
-                if len(identifier_in_primary_source) == 1:
-                    identifier_in_primary_source = str(identifier_in_primary_source[0])
-                else:
-                    raise ValueError(
-                        f"Expected one value for identifierInPrimarySource, "
-                        f"got {len(identifier_in_primary_source)}"
-                    )
-            else:
-                identifier_in_primary_source = str(identifier_in_primary_source)
-        else:
-            raise ValueError("Missing value for `identifierInPrimarySource`.")
+        return identifier_type(
+            get_provider()
+            .assign(self.hadPrimarySource, self.identifierInPrimarySource)
+            .identifier
+        )
 
-        if had_primary_source := values.get("hadPrimarySource"):
-            if isinstance(had_primary_source, list):
-                if len(had_primary_source) == 1:
-                    had_primary_source = MergedPrimarySourceIdentifier(
-                        had_primary_source[0]
-                    )
-                else:
-                    raise ValueError(
-                        f"Expected one value for hadPrimarySource, "
-                        f"got {len(had_primary_source)}"
-                    )
-            else:
-                had_primary_source = MergedPrimarySourceIdentifier(had_primary_source)
-        else:
-            raise ValueError("Missing value for `hadPrimarySource`.")
+    def _get_stable_target_id(
+        self, identifier_type: type[_MergedIdentifierT]
+    ) -> _MergedIdentifierT:
+        """Consult the identity provider to get the `stableTargetId` for this item.
 
-        provider = get_provider()
-        identity = provider.assign(had_primary_source, identifier_in_primary_source)
+        Args:
+            identifier_type: MergedIdentifier-subclass to cast the identifier to
 
-        # In case an identity was already found and it differs from the identifier
-        # provided to the constructor, we raise an error because it should not be
-        # allowed to change the identifier of an existing item.
-        if identifier := values.get("identifier"):
-            if isinstance(identifier, list):
-                if len(identifier) == 1:
-                    identifier = identifier[0]
-                else:
-                    raise ValueError(
-                        f"Expected one value for Identifier, got {len(identifier)}"
-                    )
-            if identity.identifier != str(identifier):
-                raise ValueError("Identifier cannot be set manually to new value.")
+        Returns:
+            StableTargetId of the correct type
+        """
+        from mex.common.identity import get_provider  # break import cycle, sigh
 
-        # In case an identity was found, we allow assigning a new stable target ID
-        # for the purpose of merging two items, except for the MEx
-        # primary source itself.
-        if stable_target_id := values.get("stableTargetId"):
-            if isinstance(stable_target_id, list):
-                if len(stable_target_id) == 1:
-                    stable_target_id = stable_target_id[0]
-                else:
-                    raise ValueError(
-                        f"Expected one value for stableTargetId, "
-                        f"got {len(stable_target_id)}"
-                    )
-            if (
-                identity.stableTargetId != str(stable_target_id)
-                and stable_target_id != MEX_PRIMARY_SOURCE_STABLE_TARGET_ID
-            ):
-                raise ValueError(
-                    "Cannot change `stableTargetId` of MEx primary source."
-                )
-
-        # update instance values
-        values["identifier"] = identity.identifier
-        values["stableTargetId"] = identity.stableTargetId
-        return values
+        return identifier_type(
+            get_provider()
+            .assign(self.hadPrimarySource, self.identifierInPrimarySource)
+            .stableTargetId
+        )