+++ mex/common/primary_source/transform.py
@@ -2,7 +2,6 @@ from collections.abc import Generator, Iterable
 
 from mex.common.logging import watch
 from mex.common.models import (
-    MEX_PRIMARY_SOURCE_IDENTIFIER_IN_PRIMARY_SOURCE,
     MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
     ExtractedPrimarySource,
 )
@@ -22,17 +21,10 @@ def transform_seed_primary_sources_to_extracted_primary_sources(
         Generator for ExtractedPrimarySource
     """
     for primary_source in primary_sources:
-        if primary_source.identifier == MEX_PRIMARY_SOURCE_IDENTIFIER_IN_PRIMARY_SOURCE:
-            set_stable_target_id = dict(
-                stableTargetId=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID
-            )
-        else:
-            set_stable_target_id = dict()
         yield ExtractedPrimarySource(
             identifierInPrimarySource=primary_source.identifier,
             title=primary_source.title,
             hadPrimarySource=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
-            **set_stable_target_id,
         )