+++ tests/wikidata/test_extract.py
@@ -1,9 +1,11 @@
 import json
+from typing import Any
 
 import pytest
 from pytest import MonkeyPatch
 
 from mex.common.exceptions import MExError
+from mex.common.types import TextLanguage
 from mex.common.wikidata.connector import (
     WikidataAPIConnector,
     WikidataQueryServiceConnector,
@@ -11,6 +13,7 @@ from mex.common.wikidata.connector import (
 from mex.common.wikidata.extract import (
     _get_organization_details,
     search_organization_by_label,
+    search_organizations_by_label,
 )
 from tests.wikidata.conftest import TESTDATA_DIR
 
@@ -23,6 +26,27 @@ def test_search_organization_by_label() -> None:
 
     assert search_result, f"No organizations were found for id: {expected}"
     assert search_result.identifier == expected
+
+
+@pytest.mark.integration
+def test_search_organizations_by_label() -> None:
+    search_result = list(
+        search_organizations_by_label(
+            item_label='Robert Koch Institute"',
+            offset=0,
+            limit=10,
+            lang=TextLanguage.EN,
+        )
+    )
+    identifier = "Q679041"
+    labels = {
+        "de": {"language": "de", "value": "Robert Koch-Institut"},
+        "en": {"language": "en", "value": "Robert Koch Institute"},
+    }
+
+    assert len(search_result) == 3
+    assert search_result[0].identifier == identifier
+    assert search_result[0].labels.model_dump() == labels
 
 
 @pytest.mark.integration
@@ -41,7 +65,7 @@ def test_search_organization_by_label_mocked_error(monkeypatch: MonkeyPatch) -> 
         },
     ]
 
-    def mocked_query_response():
+    def mocked_query_response() -> list[dict[str, dict[str, str]]]:
         return expected_query_response
 
     monkeypatch.setattr(
@@ -164,7 +188,7 @@ def test_search_organization_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
         },
     ]
 
-    def mocked_query_response():
+    def mocked_query_response() -> list[dict[str, dict[str, str]]]:
         return expected_query_response
 
     monkeypatch.setattr(
@@ -173,7 +197,7 @@ def test_search_organization_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
         lambda self, _: mocked_query_response(),
     )
 
-    def mocked_item_details_response():
+    def mocked_item_details_response() -> Any:
         with open(TESTDATA_DIR / "items_details.json", encoding="utf-8") as f:
             data = json.load(f)
             return data[0]
@@ -203,7 +227,7 @@ def test_search_organization_by_label_for_none_mocked(monkeypatch: MonkeyPatch) 
         },
     ]
 
-    def mocked_query_response():
+    def mocked_query_response() -> list[dict[str, dict[str, str]]]:
         return expected_query_response
 
     monkeypatch.setattr(
@@ -212,7 +236,7 @@ def test_search_organization_by_label_for_none_mocked(monkeypatch: MonkeyPatch) 
         lambda self, _: mocked_query_response(),
     )
 
-    def mocked_item_details_response():
+    def mocked_item_details_response() -> Any:
         with open(TESTDATA_DIR / "items_details.json", encoding="utf-8") as f:
             data = json.load(f)
             return data[0]
@@ -226,6 +250,146 @@ def test_search_organization_by_label_for_none_mocked(monkeypatch: MonkeyPatch) 
     search_result = search_organization_by_label(item_label="TEST")
 
     assert search_result is None
+
+
+@pytest.mark.usefixtures(
+    "mocked_session_wikidata_query_service", "mocked_session_wikidata_api"
+)
+def test_search_organizations_by_label_mocked(monkeypatch: MonkeyPatch) -> None:
+    expected_query_response = [
+        {"item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q26678"}},
+        {"item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q821937"}},
+        {"item": {"type": "uri", "value": "http://www.wikidata.org/entity/Q125698083"}},
+    ]
+
+    expected_organization = {
+        "identifier": "Q26678",
+        "labels": {
+            "de": {"language": "de", "value": "TEST"},
+            "en": {"language": "en", "value": "TEST"},
+        },
+        "claims": {
+            "website": [
+                {
+                    "mainsnak": {
+                        "datavalue": {
+                            "value": {
+                                "text": "https://www.testgroup.com/",
+                                "language": None,
+                            }
+                        }
+                    }
+                },
+                {
+                    "mainsnak": {
+                        "datavalue": {
+                            "value": {"text": "https://www.test.com", "language": None}
+                        }
+                    }
+                },
+            ],
+            "isni_id": [
+                {
+                    "mainsnak": {
+                        "datavalue": {
+                            "value": {"text": "0000 0001 2308 257X", "language": None}
+                        }
+                    }
+                }
+            ],
+            "ror_id": [
+                {
+                    "mainsnak": {
+                        "datavalue": {"value": {"text": "05vs9tj88", "language": None}}
+                    }
+                },
+                {
+                    "mainsnak": {
+                        "datavalue": {"value": {"text": "044kkbh92", "language": None}}
+                    }
+                },
+            ],
+            "official_name": [
+                {
+                    "mainsnak": {
+                        "datavalue": {
+                            "value": {
+                                "text": "Bayerische Motoren Werke AG",
+                                "language": "de",
+                            }
+                        }
+                    }
+                }
+            ],
+            "short_name": [],
+            "native_label": [
+                {
+                    "mainsnak": {
+                        "datavalue": {
+                            "value": {"text": "test_native_name", "language": "de"}
+                        }
+                    }
+                }
+            ],
+            "gepris_id": [],
+            "gnd_id": [
+                {
+                    "mainsnak": {
+                        "datavalue": {"value": {"text": "2005475-0", "language": None}}
+                    }
+                }
+            ],
+            "viaf_id": [
+                {
+                    "mainsnak": {
+                        "datavalue": {"value": {"text": "129013645", "language": None}}
+                    }
+                }
+            ],
+        },
+        "aliases": {
+            "de": [
+                {"language": "de", "value": "alias_de_1"},
+                {"language": "de", "value": "alias_de_2"},
+                {"language": "de", "value": "alias_de_3"},
+            ],
+            "en": [
+                {"language": "en", "value": "alias_en_1"},
+                {"language": "en", "value": "alias_en_2"},
+                {"language": "en", "value": "alias_en_3"},
+                {"language": "en", "value": "alias_en_4"},
+            ],
+        },
+    }
+
+    def mocked_query_response() -> list[dict[str, dict[str, str]]]:
+        return expected_query_response
+
+    monkeypatch.setattr(
+        WikidataQueryServiceConnector,
+        "get_data_by_query",
+        lambda self, _: mocked_query_response(),
+    )
+
+    def mocked_item_details_response() -> Any:
+        with open(TESTDATA_DIR / "items_details.json", encoding="utf-8") as f:
+            data = json.load(f)
+            return data[0]
+
+    monkeypatch.setattr(
+        WikidataAPIConnector,
+        "get_wikidata_item_details_by_id",
+        lambda self, _: mocked_item_details_response(),
+    )
+
+    search_result = list(
+        search_organizations_by_label(
+            item_label="TEST", offset=0, limit=10, lang=TextLanguage.EN
+        )
+    )
+
+    assert len(search_result) == 3
+    assert search_result[0].model_dump() == expected_organization
 
 
 @pytest.mark.integration
@@ -430,7 +594,7 @@ def test_get_organization_details_mocked(monkeypatch: MonkeyPatch) -> None:
         },
     }
 
-    def mocked_item_details_response():
+    def mocked_item_details_response() -> Any:
         with open(TESTDATA_DIR / "items_details.json", encoding="utf-8") as f:
             data = json.load(f)
             return data[0]