+++ CHANGELOG.md
@@ -9,15 +9,33 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add validator to base model that verifies computed fields can be set but not altered
+- new class hierarchy for identifiers: ExtractedIdentifier and MergedIdentifier
+
 ### Changes
+
+- improve typing for methods using `Self`
+- make local type variables private
+- use json instead of pickle to calculate checksum of models
+- replace `set_identifiers` validator with computed fields on each extracted model
 
 ### Deprecated
 
 ### Removed
 
+- removed custom stringify method on base entities that included the `identifier` field
+
 ### Fixed
 
+- fix typing for `__eq__` arguments
+
 ### Security
+
+## [0.28.0] - 2024-07-08
+
+### Added
+
+- extract multiple organizations from wikidata
 
 ## [0.27.1] - 2024-06-14